Update movies to allow for appended data.
I like this version better. Instead of "hard coding" a range of 5, I made the range just the length of the smallest list.  Now, if one chooses to append data to the lists, that new data will be output as well. But, if you're appended data haphazardly, and don't include all the data, it will only print the data for the length of the shortest list.  I think that's a more appropriate approach for reusability.