Create states
I did this a little differently, since I started with all of the states and abbreviations from the USPS postal abbreviation website. Those are all in upper case, and I wanted the result to have proper casing.

I think I'm going to see if I can loop the casing for each word in the name of a state.  That would be more elegant, so it's just that I happen to know that there's no state with more than three words.